+++ CHANGELOG.md
@@ -5,12 +5,17 @@ All notable changes to this project will be documented in this file.
 ## [v0.4.3] - 2024-08-05
 
 ### Added
+
+- Added filler items:
+    - Invincibility, 5 Rings, 10 rings, Shield, Magnetic Shield, Extra life.
 - Added Gamma's Hot shelter 4th life capsule
+- Added error message when generating without any character.
 
 ### Changed
 
 - Fixed death link and ring link not working together.
 - Fixed crash when receiving a death link while in the main menu.
+- Fixed receiving extra items after reconnecting to the server.
 
 ## [v0.4.2] - 2024-08-04